Fix allocator_startArc_particleException_failsWaitForStart in AllocatorTestBase.
There's a platform-specific check required for this, due to the returned exception being passed directly back (for CoreAllocatorTest) or through a round of serialization (for AndroidAllocatorTest). I've handled this directly in AllocatorTestBase because I think it's better to keep the logic in one place, rather than putting a tiny check-exception-for-this-one-test method to be overridden in the two test classes.