Encode Type and Schema to proto
A lot of proto decoding functionality already exists, but most encoding functionality is missing. We're going to need to be able to encode lots of things to proto in order to serialise PolicyConstraints properly. This is the first step in adding that proto encoding, starting at the bottom of the stack (types) and working upwards.