Cleanup coroutine usage with Stores
* Remove the `coroutineContext` parameter from the Reference interface
methods. These are already suspend functions, so the caller should
switch to that context before invoking the function. I didn't notice any
use cases that required this.  The call in `dereference` to
`store.onProxyMessage(SyncRequest)` was being launch in a coroutine on
the provided context, but all tests pass running this method directly on
the current coroutine.

* Remove the `coroutineContext` property from `StoreOptions`. As best I
can tell, this was not actually being used, except by the
`cleanCoroutineContext` parameter of `ReferenceModeStore`.

* Initialize the `ReferenceModeStore` with the `coroutineContext` on
which construction occurs, to be used for the cleanup flow. Added a test
to ensure that the flow performs as expected.