Refactor RefModeStore cleanup behavior
This set of changes began as a result of the tests added here, that
failed.

This refactor makes two improvements:
1. Fixes an issue where once cleanup was triggered for a store, it would
never be triggered again.
2. No need to launch a separate coroutine for cleanup, or to provide any
coroutineContext at construction time. Cleanup now occurs on the
coroutine that resulted in the internal state change that triggered
cleanup.

In order to enable cleanup to always occur on a caller coroutine, the
`ActiveStore.off` method needed to become a suspend method. This allows
us to post a message to the `MessageQueue` when `off` is called, to
ensure that cleanup occurs if `off` was called while the `MessageQueue`
was already idle.

Luckily, this `suspend` change didn't cause many downstream changes;
most calls to `off` were already in a coroutineContext. The one
exception to this was `storageProxy.close`. So this is now also a
suspending method.

An additional change here was to remove the use of a coroutine mutex in
ProxyCallbackManager, in favor of a symbol object monitor, since most of
the methods interacting with the component aren't suspendable, anyway.