Do we even need a receive queue?
While investigating b/159144117, I made the following observations:

* We currently may have multiple receive queue batches draining
concurrently. The bug referenced above states that we need to fix that,
and only process one receive queue operation at a time.

* Calls to `enqueue` on the receiver queue are synchronous; the call to
enqueue will not return until the message has been processed by the
queue.

So, why don't we just synchronously call the methods that were being
placed on the queue?

So far, this seems to work: all tests are passing.