Simplify ReferenceModeStore queue implementation
Alternative solution for b/159144117.

Replaces the custom MessageQueue and SendQueue with a single, simpler
SimpleQueue, which is just a basic serialized operation queue that runs
suspend functions one by one.

The HoldQueue is still present; we work with it directly in
ReferenceModeStore now. It's not really a queue, just a map of pending
operations. We pass to it the send queue for running its callback jobs.

There are two major pieces to consider here when reviewing:
1. The queue itself.
2. The `OperationQueue` interface, and the changes in
`ReferenceModeStore` that switch to using this interface for the send
and receive queues.

This CL fixes a few issues as well:
* Receive queue operation could potentially start multiple concurrent
queue drains. This should no longer be possible.
* The cleanup job would only run once. It should now run every time the
callbacks count becomes 0 and the queue becomes empty.