Backing store data class: Default to schema hash when `name` is null. (#5745)
Fixes issue with unnamed schemas, type variables being a common use case. A data class for backing stores in storage key factories initializes a field in such a way that null values are possible. Instead of choosing a null value, this PR will make a field on the data class default to the Schema's `hash` value.