Proto encoding for Type, FieldType, Schema (#5709)
* Rename TypeProtoDecoders.kt to TypeProto.kt, since it is about to get some encoder functions

* Encode Type and Schema to proto

A lot of proto decoding functionality already exists, but most encoding functionality is missing. We're going to need to be able to encode lots of things to proto in order to serialise PolicyConstraints properly. This is the first step in adding that proto encoding, starting at the bottom of the stack (types) and working upwards.

* Missing changes to SchemaProto.kt

* Re-renamed files