Allow inline declarations to reference schemas that are already defined (#5781)
The aim is to get things like this working:

```
schema ExternallyDefined
  ...

schema LookAtMe 
  field: inline ExternallyDefined
```

It turns out this is an extremely small change because the processing of TypeName in manifest.ts is universal (we always replace a TypeName with the actual schema, regardless of the position in the AST). Sometimes we can have Nice Things!