Fixes bug with parsing of DFA field paths and type variables.
The field path validation code currently only works against unresolved type variables (i.e. type variables with constraints). However, when that type variable gets fully resolved, the type constraints are lost. (To repro this bug, call `normalize()` and `clone()` on a recipe with a type variable.)

The fix is to fallback to the type variable's resolved type if the constraint is missing. This should be safe, since the original ParticleSpec from the manifest is known to have already passed validation, meaning the resolved type (a superset of the constraints) should also pass.

This PR includes another fix, where resolved types were not being copied across to the new ParticleSpec until after the constructor had already run (and the constructor is where the validation code was invoked, hence this would fail).