Change Handle.onUpdate to receive a delta object instead of the model. (#5830)
The onUpdate callback currently receives the result of `handle.fetch()` or `.fetchAll()` as its argument. This is replaced by a `SingletonDelta` or `CollectionDelta` object that indicates how the handle's data has changed (the singleton case is just the old and new entities; we may make that more detailed in the future).