Introduce the new `StorageEndpointManager` interface
The `StorageEndpointManager` interface serves out
`StorageEndpointProviders`, which could be thin wrappers around
`ActiveStores`, but could also be something that communicates over
process or network boundaries to something else that's managing stores
(for example, an Android service).

This helps us to build a distinction between a `Store` and a
`StorageEndpoint`.

This PR also removes the storage endpoint generation code from
`ActiveStore` and puts it in the provider implementation, to further
cement the different between a `Store` and ` StorageEndpoint`.

In this PR, we draw a boundary for changes in `EntityHandleManager`, by
providing an adapter method that converts a `StoreManager` into a
`StorageEndpointProvider.

Follow-up work will remove `StoreManager` and `ActivationFactory`
completely, instead providing concrete `StorageEndpointProvider`
implementations for Android and vanilla JVM.