Remove `lifecycle` from `ServiceStore` API (#5800)
A lot has changed since ServiceStore was originally built, and this
feature doesn't make as much sense now (and is causing difficulties).

The primary reasons to move forward on removing this now are:
1) LifecycleRegistry is not thread safe. In a recent update, it now also
enforces main thread interaction.

2) AiAi's current usage patterns are not owned by a lifecycle at all, so
we have to synthetically create one, anyway.

We're only using the lifecycle functionality to trigger cleanup on
destroy, so we don't lose a lot by removing this. But we need to take
care in some contexts to make sure to clean up any owned stores.