Allow multiple instances of particles in recipes (#5818)
Our plan & serialization formats support multiple instances of the same particle, but prior to this fix we would collapse these list representations into a map representation at runtime, which would therefore conflate those instances.

Note that there are currently no new tests demonstrating that this fix works; I'd love some suggestions on where to add them. Additionally, the pending inlines showcase will make use of this functionality.

There's likely further work to be done here - if there are multiple instances of a particle that need to be accessed from outside the context of an arc, then we need a way to disambiguate between them (or possibly just a way to return all instances matching a particular particle name).