Make the registerCallback and unregisterCallback calls in ServiceStore non-blocking
As part of this, the old `DeferredResult` helper class is removed,
replaced by the `suspendForCallback` family of helper methods. When
interacting with a service object that has a single callback.

Now that all of the service methods are oneway/async, and the on/off
methods of [ServiceStore] are suspend, we can now consistently use this
suspendForCallback-based pattern everywhere that we are calling into the
service.