Remove the channel from `ServiceStore` (#5796)
It doesn't appear to be providing us benefits anymore, as the
ServiceStore code has evolved:

`onProxyMessage` and `off` were queueing work onto the channel. It seems
this was to prevent concurrent runs of onProxyMessage but since that
method is markedi `oneway`, and the result await did not block channel
execution, I'm not sure it was providing that guarantee.

All existing tests pass with this change; if there's a case that this
was addressing that I haven't considered, then we need a test for it.