Update PolicyVerifier to use the new `@policy` recipe annotation. (#5864)
* Update PolicyVerifier to use the new `@policy` recipe annotation.

This new annotation dictates which policy should apply to the recipe. Trying to use any other policy with the recipe will be a failure.

This also replaces the `PolicyOptions.allowedEgresses` field, which was used to indicate which egress particles are allowed by a policy. Instead, we will allow any egress particle whose egress type matches the egress type specified in the policy.

* dep

* review

* lint