Delete manifestChecker / arcs_manifest_parse_test
It was only used in a couple of places, and isn't actually very helpful. This script would check that a .arcs file parsed correctly, and that the resources it mentioned (e.g. other .arcs files, .js/.json/.wasm files) actually exist.

However, most of those things we aren't using right now, and we have other BUILD rules that check a manifest file parses correctly as a side-effect (e.g. arcs_manifest_proto). Additionally, manifestChecker was too strict about certain things (e.g. it checks that a particle spec has either an implementation file or is marked `external`) which we don't want to enforce right now.

So, I'm deleting it. It can be ressurected in the future if we decide that it was actually helpful.