Remove `runBlocking` from DirectStore
This means that we can no longer directly close stores as part of
LruCacheMap eviction, since the get/put methods of the map are not
suspendable.

I tried one approach where I just made a parallel implementation of
LruCache map with suspendable get/put. That worked ok, but was a lot of
code duplication.

In this approach, I try queueing the stores for closure by the suspend
method that's creating them.