Construct AbstractArcHost with a HandleManagerProvider
This is a simpification of the AbstractArcHost interface. Rather than
 providing the various pieces needed to create handles (Time, Stores,
Schedulers, etc) you provide a HandleManagerProvider than encapsulates
all of these things.

* removes the `platformTime` overridable val
* removes the `stores` overrideable val
* removes the `entityHandleManager` overrideable method

In this PR, we continue using the same implementation for returning
Handles under the hood, the only change here is in how things are
structured.

Once this PR is landed, we can create StorageEndpointProvider
implementations for JVM and Android, and creating matching
HandleManagerProvider implementations that use this. P