Defines a run_test helper BUILD rule and policy_test for manifest files
The run_test helper rule is there to invoke a test binary with custom args, which is apparently something that bazel does not support out of the box.

The arcs_manifest BUILD rule can now also generate a policy_test rule, which runs the VerifyPolicy program against the manifest.

Also added a testcase to exercise the new rule.