Two fixes for inline entities in Databases (#5938)
* Two fixes for inline entities in Databases

(1) ensure that inline entities get returned from the Database driver as
RawEntities rather than as References. We were incorrectly building
references for them rather than wrapping as Referencables due to a
missing case in the Database driver.
(2) ensure that inline entity IDs are returned from the Database driver.
We were previously not fetching these as an extra query is required;
however this meant that collections were deduping all but one entity on
return from the DB.

Also added some simple end-to-end store-and-retrieve tests in the new
StorageServiceEndToEndTest file.