Clarify StorageKeyParser interface, make global instance thread-safe
This introduces an official interface for StorageKeyParsers, and uses a
consistent pattern for defining them in the storage key classes.

Along with this, the current “StorageKeyParser” top-level object now
implements a StorageKeyManager interface, which describes an object that
can delegate parsing to a collection of registered parsers.

The global manager object’s internal map is now accessed in a
thread-safe manner.

b/164161757