If a protected store is not marked, policy verifier rejects recipes. (#5934)
Currently, we assume that all stores that need to be protected by a
policy are present in the `storeClaims` map of the `PolicyConstraints`
data structure. However, it is possible we may accidentally not
identify all the protected stores (if a store is not named correctly
in the recipe, for instance). If that happens, the policy verifier
should do the safest action, which would be to reject the recipes.

To do so, if a store corresponding to a handle is not in
`storeClaims`, we consider that the store has data that cannot be
accessed if any of the following holds:
   - handle has only read connections.
   - handle is written by a particle that only has write connections.

Both the above conditions indicate that the data was possibly
ingested into Arcs from outside and should have been protected.