Ensure inline entities are also deleted when a parent entity is. (#5964)
Prior to this patch, inline entities were not being removed when
a parent entity was rewritten, nor were they being removed as
part of garbage collection.

Both these processes make use of delete() to remove the parent
entity. This patch allows delete() to renmove inline entities as
well.

How this works:

Inline entity storage keys are of the pattern
  inline://{<parent key>}<extra data>

Hence, when a storage key is deleted using delete(), we now
look for that exact storage key, as well as any storage keys
that are LIKE "inline://{%storageKey}%". This will match
inline keys that have the parent key somewhere in their
chain of parent keys.