Adds a new `@ingress` particle annotation
This will be used to indicate that a particle ingresses data from outside Arcs. It will eventually be used to replace the PolicyOptions object (but not in this PR).

I had to rework the representation of the isolated/egress properties, since there is now a valid combination state (ingress+egress).