Expand the `StoreEndpointManager` pattern outward (#5979)
This is another step in the direction of creating a class structure
thatmaintains a distinction between an `ActiveStore` and a single
channel to that store (`StorageEndpoint`).

Now `EntityHandleManagers` must be provided a `StorageEndpointManager`
at construction time.

This introduces the implementation of a `DirectStorageEndpointManager`,
which provides endpoints directly wrapping an `ActiveStore`. For now,
that's also providing Android storage endpoints, because those are still
trying to behave like `ActiveStore`s. But this will change.

Since the `EntityHandleManager` now gets a `StorageEndpointManager`
instead of an `ActivationFactory` (those are on their way out), the
dereferencing path is also update to work with these interfaces as well.

No major behavior change here, just naming and interfaces.

`StorageEndpoint` now implements the `Closeable` interface, so it can be
used with the `use` helper in Kotlin, we use this to our advantage in
the `RawEntityDereference`.