Start moving recipe/ dependencies to a new lib-recipe.ts.
This is getting us ready to move lib-recipe into arcs-types and keep
internal recipe details independent of the rest of the codebase.

There are 3 activities here; all need to be completed before we are
modular / in a good place:
(1) make all non-recipe code that depends on recipe/ do so through
lib-recipe.ts (e.g. this patch)
(2) make sure recipe/ files only depend on things inside recipe/, or
arc-types.
(3) rationalize the lib-recipe.ts API.