Make `on` method suspending, option 1
Having the `on` method suspend instead of block opens up the possibility
of restructuring the places that we use it to avoid deadlocks.

StorageProxy construction is fixed by making the store channel late
init, and introducing a suspending construction method that creates the
attached storage channel before returning the object.

It was also necessary to restructure `ReferenceModeStore` a bit so the
containerStore callback attachment happens in the suspendable
constructor method, as well.