Return to runBlocking for callback registration (#6009)
Blocking the caller of registerCallback can lead to a deadlock, if the
caller of registerCallback is using the same (limited) threadpool as the
BindingContext scope.

When we use `runBlocking` with no parameters, an event loop is created
on the current thread. This prevents the problem where a `scope.launch`
attempts to run a block, but no threads are available.