Improvements to `runOnAllDatabases`: run all jobs & test
This changes the behavior of `runOnAllDatabases` so that it runs all
jobs regardless of the failure of any other. Failures are returned as a
composite exception.

Questions are often raised about the behavior of `coroutineScope`. This
PR introduces a test for the `runOnAllDatabases` method to verify that
`coroutineScope` is actually behaving in the way we want.

Since the method isn't specific to Android database manager, this PR
also moves it to become an extension method on `DatabaseManager`.

Since the pattern may be useful in other contexts, I created
`CompositeException` and some generalized helpers that perform this
pattern.

I tried running the test with `coroutineScope` replaced with
`CoroutineScope(coroutineContext).launch` just to verify that the added
tests fail in that case.