Remove boolean return value from `onProxyMessage` (#6084)
Nothing meaningfully acts on this return value, and we should not
actually need it. The typescript implementation doesn't include it.

We'd already gone part of the way down this road in removing the value
for the StorageService APIs, this removes it from the `ActiveStore`
onProxyMessage definition.

The one place the value seems to be used meaningfully is in
ReferenceModeStore, when interacting with the DirectStores of the
DirectStoreMuxer. But all that happens there is throwing an exception.
That behavior isn't needed, and no tests even verified it.