Make the `off` method of stores suspending. (#6071)
This allows us to perform suspendable operations in the `off` method
without requiring runBlocking.

Two small changes were needed:
* We can no longer use `use` with `StorageEndpoint`s, since `suspend
close` doesn't meet the `Closable` interface.

* Removing runblocking from the `off` method of ReferenceModeStore
changes the behavior of that method a bit; the empty method is not
guaranteed to run by the time the method exists. This resulted in some
test failures. To address it, the `idle` method now also makes sure that
the receiveQueue is idle.