Improve type tracking through recipe.updateToClone. (#6120)
This used to be typed as returning any; now it returns the same type
it is given, which means the output dict members all automatically
have the correct type.

It does make generic uses slightly more painful (e.g. see
convert-constraints-to-connections.ts) but even here we've now got
more type safety as we're forced to declare what we expect each output
to be.