Fix a race condition during particle shutdown
During particle shutdown, there's a window in which we'd like to stop
receiving storage events for the handle, but we'd like to allow handles to
continue to use the handles to write (while executing the on Shutdown
method).

To allow this, handles now have a "detach" method, and handle holders
will:
* detach handles
* call onShutdown
* close handles

These changes fixes a flakey allocator test which was sometimes receiving
storage events after removing all of the handles for the particle,
leading to a crash.