Remove the result callback from Storage Service's `sendProxyMessage`
When the StorageService was created, `onProxyMessage` calls on stores
would return a boolean result indicating success or failure. That's
since been removed.

We also removed the reporting of any meaningful information other than
an ACK that the service received the call.

Since the callback no longer provides any sort of actionable signal, we
can just remove it altogether.