Make VolatileMemory methods suspending and extract VolatileMemory interface.
This sets the stage for being able to implement a controlled-latency RamDisk or Volatile memory storage system to help sus out race conditions in Arcs storage stack components by providing a custom implementation of VolatileMemory/RamDisk which is capable of awaiting on countdown latches or using delays to emulate slow storage media.

A follow-up CL will introduce a testonly implementation of the VolatileMemory interface that provides exactly this functionality.

PiperOrigin-RevId: 333619257