Fix a very rare race condition when shutting down StorageProxy
If the scheduler associated with a storage proxy is cancelled while messages
are in flight, the flow's `onEach` operation that eventually updates the `busySendingMessagesChannel`, allowing idle to return to false, will never be run.

To resolve this, when the flow that controls `busySendingMessagesChannel`
completes, it also sends a final `false` value to that channel.

This CL adds a test which triggers the condition before the fix, and passees
with the fix applied.

PiperOrigin-RevId: 333637877