Remove the now-obsolete ActivationFactory type.
The concept of ActivationFactory, (which created different ActiveStore
implementations for different configurations) has been replaced by
StorageEndpointManager, which provides a communications channel to an
ActiveStore.

Now that we've updated all of our usages of ActivationFactory, we can remove
this type, as well.

The former DefaultActivationFactory has been renamed to Store, reflecting that
it's a constructor for an ActiveStore type.

PiperOrigin-RevId: 334851501