Provide CoroutineScope as separate DirectStore parameter, not part of StoreInfo
A CoroutineScope parameter was added to `StoreOptions` some time ago, but it
shouldn't be there: StoreOptions is for information directly related to the
data type of a store, not for parameters that are related to the particular
implementation details.

A lot of tests were updated to pass the coroutineScope, which is now a required
parameter. A new helper was added to create a StoreManager for testing, using a
new scope based on Dispatchers.Default.

Eventually, we'll want to refactor these tests so that they all properly plumb
through the coroutineScope of their test block.

PiperOrigin-RevId: 334906799