Remove global WriteBack configuration objects in favor of dependency-passing
When the WriteBack feature was introduced, some architectural shortcuts were
taken so that the feature could be added in time to meet a deadline. Now that
we've done some other cleanup related to store creation and management, it's
easier to provide a WriteBackProvider to DirectStore, DirectStoreMuxer, and
ReferenceModeStore.

This also removes the need for the testing implementation; a test WriteBack was
just the normal implementation called with different constructor parameters.

PiperOrigin-RevId: 334915454