Plumbs Storage Keys to DevTools UI (#6316)
Adding storage keys pertaining to the proxy messages required some refactoring. Concretely:

Reuse of `IStorageServiceCallback` for DevTools has been dropped for the dedicated callback interface `IDevToolsProxyCallback`that contains the storage key. This is considered a good change, as it allows to customize the protocol suitable to the DevTools task - as needed here.

`DevToolsProxy` interface exposed to the Storage Stack has been reshaped into a `DevToolsForStorage` and a set of (currently 2) slices of the API, one for the `DirectStore`, one for the `ReferenceModeStore`. This allows us to keep the metadata about the object sending/receiving the message (here: storage key) without the need for the object under inspection to hold this data (as today stores don't hold their storage keys). It also paves the way to have in the future a disjoint slice of the DevTools API that is dedicated to an ArcHost (`DevToolsForArcHost` or something like that).

Closes #6316

COPYBARA_INTEGRATE_REVIEW=https://github.com/polymerlabs/arcs/pull/6316 from piotrswigon:devtools-storage-keys 21bc217e9155860aa1247b62a51b78a673646764
PiperOrigin-RevId: 335445607