Prevent DB creation race by always running e2e test StorageService in main process
Our DatabaseImpl is not robust to multiple processes racing to run
`onCreate`, which leads to the e2e test occassionally failing with an
error about duplicate table creation.

Our current workmanager task implementations work directly with the
database. Workmanager tasks always run in the main process. This means
that, for the time being, in order to avoid this race, we need to make
the StorageService run in the main process, as well.

This test will still be exercising the cross-process AIDL binding
because it creates *another* service in a different process that binds
to the storage service.

b/170119674