Defines an AIDL interface for IMuxedStorageService, and provides the shell for the implementation.
This service is for proxying MuxedStorageProxyMessages to/from the storage service. It doesn't follow the old StorageService's format for registering callbacks, etc. and instead shows a new path forward involving "channel" objects, i.e. a way of communicating with a store (or DirectStoreMuxer in this case) without binding to it every time.

Will migrate the old service to this new format too if it looks good.

PiperOrigin-RevId: 335518145