Introduce `DriverFactory` interface
This is a step towards removing the global `DriverFactory` object.

The name `DriverFactory` now describes an interface for a class that manages a
number of `DriverProvider` items.

The object that used to use this name is now called `DefaultDriverFactory`, and
it implements the `DriverFactory` interface.

There are no major structural changes here: DefaultDriverFactory is still
accessed directly as a global object in the same way it already was.

PiperOrigin-RevId: 335766215