Remove implicit driver registrations
This is another step towards completely removing the global DriverFactory
object.

In this CL we simplify the interface to the global object so that you can only
call `resetRegistrations` with the entire set of driver providers that you'd
like to use. Thus, the set of available drivers is always determined at one
time, in one place.

PiperOrigin-RevId: 335769535