Use a stable strategy for finding drivers registered to a factory.
Previously, DriverFactory used a set to manage driver providers. The order of
iteration for a set is indeterminate, and so in a case where multiple providers
supported the same key, you might receive a different driver in different
situations.

This changes the factory to use a list, and adds documentation noting that, in
the case of multiple drivers supporting the same key, the one you place
earliest in the list will have priority.

As part of this, a common FakeProvider is created to handle the tests added for
this behavior. It's also designed so that it can replace the one that was used
in StoreTest.

PiperOrigin-RevId: 336716233