Introduce a new pattern for connecting to service in Arcs
Our needs have outgrown the capabilities of the current service connection
helpers, resulting in some oddly shapped patterns, and some runtime checks.

In this change, we introduce a new foundational service connection helper that
pulls out the useful pattern that the StorageServiceConnection provides,
updates it with our updated coroutines knowledge.

It also separates the functionality related to platform-specific binding
strategies from the details about creating Intents.

With this change, we can switch our two non-StorageService connections
(DevTools and StorageServiceManager) to use this.

Since we're no longer using the ConnectionFactory/BindingDelegate pairs for
non-StorageService bindings, the storeOptions parameter is returned to being
non-nullable, removing the need for quite a few runtime checks.

In a subsequent change, we will update AndroidStorageServiceEndpointManager to use
this pattern, and remove StorageServiceConnection completely.

PiperOrigin-RevId: 337166401