Switch to BoundService connection method for StorageService, remove StorageServiceConnection
The BoundService pattern was introduced in cl/337166401, but in the interest of
keep changes scoped, AndroidStorageServiceEndpointManager was not converted at
that time.

In this CL we make that conversion, allowing us to remove
StorageServiceConnection, ConnectionFactory, and BindingDelegate.

Helper methods are included in StorageService to allow overriding the class
when creating a StorageService binding intent; it's no longer necessary to
reimplement that method in your own class.

Also includes a fix to allow this pattern to work for AIDL services as well.

PiperOrigin-RevId: 337580992