Provide driverFactory as a parameter to store objects.
This is another step towards eventually removing the global
DefaultDriverFactory completely. In this CL, we pass a driverFactory to some
classes as a constructor parameter, rather than having them reference
DefaultDriverFactory directly.

Now DefaultDriverFactory.get() is only used in two places: StorageService, and
StoreManager (which plays a role similar to StorageService).

PiperOrigin-RevId: 337965008