Clean up DirectStorageEndpointManager, rename to LocalStorageEndpointManager
The original DirectStorageEndpointManager implementation was transitional; a
way to provide something conforming to the StorageEndpointManager interface
while still using existing pieces underneath (like StoreManager).

In this CL, we clean up the implementation of today's
DirectStorageEndpointManager, removing the separate StoreManager abstraction
and managing the stores directly in the manager.

To help mark the difference, and to better reflect the nature of the
implementation, it's renamed to LocalStorageEndpointManager.

Other notes:
* Introducing testDatabaseStorageEndpointManager as a test convenience
* Remove now-unneeded waitForIdle/resets in a number of test teardowns

PiperOrigin-RevId: 338291847