Remove non-operational logic in StoreWriteBack flow's onCompletion
There are no situation in which the channel drain logic in onCompletion will
actually run.

We can reach onCompletion because:
* The channel was cancelled. In this case, `channel.consumeEach` will have
  nothing to do.
* The channel was closed. In this case, all channel elements will have been
  consumed before we reach onCompletion.
* The scope was canceled. In this case, the suspending `consumeEach` method
  will throw an exception.

It's also worth noting that even for a proper `close` call, we use
channel.cancel which clears out the channel anyway. Scope cancellation is a
more drastic cleanup signal (we should not try to do much cleanup during the
cancellation cleanup phase. So it doesn't make sense that we were trying harder
to service remaining jobs in the channel on cancellation, but not on `close`.

PiperOrigin-RevId: 338302533