Launch bulk driver jobs concurrently and collect exceptions
Previously, the remove* methods on DriverFactory were wrapping a single launch
in a coroutineScope, running all of the jobs serially and making the wrapping
coroutineScope a bit pointless.

In this CL we take advantage of the CompositeException helper introduced a
while back to implement a similar pattern for the database manager. Now all
jobs will be run to completion regardless of the failure of any other jobs. If
there are any failures, the method will throw a composite excpetion containing
all of the exceptions that occurred; the caller can decide what to do about
these.

Most of this CL is the addition of tests to FixedDriverFactory test to verify
the behavior when calling removeAllEntities/removeEntitiesCreatedBetween. A new
CallbackChoreographer test helper is introduced, as an attempt to formalize the
two-job pattern that's been emerging in various places as a way to test timing
conditions related to callback invocation.

PiperOrigin-RevId: 339479495