In DatabaseImpl, wrap deletion and type initialization in the same transaction.
Without this, there is a remote possibility that an error occurs between the two steps and the database is left totally empty. Totally empty databases (no primitive types configured) will cause issues when trying to add new items to them (because the types cannot be found).

Additionally - this change uses `DROP TABLE` followed by `CREATE TABLE` semantics instead of `DELETE FROM` when resetting the tables. This means that any auto-incrementing columns are reset.

PiperOrigin-RevId: 340311684