Allow build flags to be overridden.
Only features which have status READY should ordinarily be overridden. Features with status NOT_READY, as the name would imply, are not ready for use and so cannot be overridden. Features with status LAUNCHED should not be overridden, but an exception is made for overriding them to be True (which is a no-op), to make the process of marking a flag as launched and cleaning up all the code easier.

PiperOrigin-RevId: 340591644