Introduce Type.toSchema tests; Move from TypeUtil to TypeUtils
This appears on the testing audit sheet as TypeUtil.kt needing coverage. Since
we already have TypeUtils.kt with a number of methods, I've moved the one
method in TypeUtil.kt there, and added the tests for the methods to
TypeUtilsTest.kt.

Since we have the EntitySchemaProviderType interface now, the method
implementation is also updated here to support any type that implements
EntitySchemaTypeProvider

PiperOrigin-RevId: 341936319