Created StorageChannelImpl (#6576)
`StorageChannelImpl` is an implementation of `IStorageChannel` that can communicate with an `ActiveStore`. This is part of the migration to use `IStorageServiceNg` for when a client wants to bind to the `StorageService` to communicate to an `ActiveStore`. There will be an implementation of `IStorageChannelNg` **(not yet implemented)** that will create a `StorageChannelImpl` upon invoking `openStorageChannel`.

Additionally:
- Created a base class (`BaseStorageChannel`) for the common functionality shared between `StorageChannelImpl` and `MuxedStorageChannelImpl`.
- NoopActiveStore for testing.

Closes #6576

COPYBARA_INTEGRATE_REVIEW=https://github.com/polymerlabs/arcs/pull/6576 from AliceLetitia:storage-service-impls 1474daf9072a1c25b7b502da63eb47a4fe9de8ed
PiperOrigin-RevId: 342755850