Close Schedulers when Arcs are stopped
As part of this, we remove HandleManager from ArcHostContext

A HandleManager contains a Scheduler, which should only be active and valid
while an Arc is actively running in an ArcHost. The ArcHostContext and
ParticleContext structures may live longer than the active running time of an
Arc. Maintaining a HandleManager as part of these would lead to needing a
number of runtime validity checks to make sure that the context has an active
Scheduler in situations where one is needed.

This introduces a new RunningArc structure that is a pair with an
ArcHostContext, as well as an EntityHandleManager. The methods in
AbstractArcHost are updated to accept a RunningArc where appropriate.

PiperOrigin-RevId: 343134443