Remove SchedulerProvider, provide scope to scheduler
Previously, the SchedulerProvider provided some non-trivial functionality: it
maintained a map of active schedulers to allow cancelling them, and it created
a parent scope for all of the schedulers created.

Over time, this behavior has been removed:
* The earlier complex thread-allocation mechanisms that some providers used
  have been removed; instead we now provide a dispatcher that can run the
  schedulers.
* The creator of a scheduler should already be cleaning up the schedulers, so
  there's no need for a cancelAll method anymore. In case of lifecycle mistakes
  related to the scheduler, they'll still be cleaned up when their owning
  component's scope is canceled.

Since the SchedulerProvider has been reduced to almost no functionality of note. This
CL removes it completely. Now, components that need to create a scheduler
provide it with a scope.

PiperOrigin-RevId: 343158559