Remove direct reference to ActiveStore in BindingContext
Because of our current storage service binding strategy, we end up with a large
number of BindingContext objects that are never cleaned up. There's work in
flight to introduce a new service binding pattern to address this.

This CL is a small fix to prevent the BindingContext from acting as a GC root
for stores. This is achieved by providing a function that grabs the ActiveStore
out of the map that StorageService maintains. So the BindingContext now
references StorageService, instead of a specific store object. This is OK,
since the StorageService is already long-lived.

PiperOrigin-RevId: 343590847