Build rule for a stopwords detection test.
The test looks at all flags that are disabled in the given build, collects the stopwords lists for each of those features, and runs the stopwords.sh script to ensure that the built apks don't contain any of the disabled features.

Not hooked up anywhere yet.

Also sets the grep -E flag to enable extended regular expressions, so we can combine regexes like so: (abc|def|ghi).

PiperOrigin-RevId: 343941058